feat(tests): Split storager tests into files
