+++ tests/storager_delete.go
@@ -1,0 +1,40 @@
+package tests
+
+import (
+	"bytes"
+	"io"
+	"math/rand"
+
+	"github.com/google/uuid"
+	"github.com/stretchr/testify/suite"
+
+	"go.beyondstorage.io/v5/pkg/randbytes"
+)
+
+type storageDeleteSuite struct {
+	suite.Suite
+
+	p *StorageSuite
+
+	size    int64
+	content []byte
+	path    string
+}
+
+func (s *storageDeleteSuite) SetupTest() {
+	var err error
+
+	s.size = rand.Int63n(4 * 1024 * 1024) // Max file size is 4MB
+	s.content, err = io.ReadAll(io.LimitReader(randbytes.NewRand(), s.size))
+	s.NoError(err)
+
+	s.path = uuid.New().String()
+
+	_, err = s.p.store.Write(s.path, bytes.NewReader(s.content), s.size)
+	s.NoError(err)
+}
+
+func (s *storageDeleteSuite) TestDelete() {
+	err := s.p.store.Delete(s.path)
+	s.NoError(err)
+}
